Added integration tests

+++ src/test/java/uk/gov/ons/ssdc/supporttool/testhelper/IntegrationTestHelper.java
@@ -18,6 +18,7 @@ import uk.gov.ons.ssdc.common.model.entity.Survey;
 import uk.gov.ons.ssdc.common.model.entity.UacQidLink;
 import uk.gov.ons.ssdc.common.model.entity.User;
 import uk.gov.ons.ssdc.common.model.entity.UserGroup;
+import uk.gov.ons.ssdc.common.model.entity.UserGroupAdmin;
 import uk.gov.ons.ssdc.common.model.entity.UserGroupAuthorisedActivityType;
 import uk.gov.ons.ssdc.common.model.entity.UserGroupMember;
 import uk.gov.ons.ssdc.common.model.entity.UserGroupPermission;
@@ -29,6 +30,7 @@ import uk.gov.ons.ssdc.supporttool.model.repository.PrintTemplateRepository;
 import uk.gov.ons.ssdc.supporttool.model.repository.SmsTemplateRepository;
 import uk.gov.ons.ssdc.supporttool.model.repository.SurveyRepository;
 import uk.gov.ons.ssdc.supporttool.model.repository.UacQidLinkRepository;
+import uk.gov.ons.ssdc.supporttool.model.repository.UserGroupAdminRepository;
 import uk.gov.ons.ssdc.supporttool.model.repository.UserGroupMemberRepository;
 import uk.gov.ons.ssdc.supporttool.model.repository.UserGroupPermissionRepository;
 import uk.gov.ons.ssdc.supporttool.model.repository.UserGroupRepository;
@@ -49,6 +51,7 @@ public class IntegrationTestHelper {
   private final UserRepository userRepository;
   private final UserGroupRepository userGroupRepository;
   private final UserGroupMemberRepository userGroupMemberRepository;
+  private final UserGroupAdminRepository userGroupAdminRepository;
   private final UserGroupPermissionRepository userGroupPermissionRepository;
 
   public IntegrationTestHelper(
@@ -61,6 +64,7 @@ public class IntegrationTestHelper {
       UserRepository userRepository,
       UserGroupRepository userGroupRepository,
       UserGroupMemberRepository userGroupMemberRepository,
+      UserGroupAdminRepository userGroupAdminRepository,
       UserGroupPermissionRepository userGroupPermissionRepository) {
     this.surveyRepository = surveyRepository;
     this.collectionExerciseRepository = collectionExerciseRepository;
@@ -71,28 +75,34 @@ public class IntegrationTestHelper {
     this.userRepository = userRepository;
     this.userGroupRepository = userGroupRepository;
     this.userGroupMemberRepository = userGroupMemberRepository;
+    this.userGroupAdminRepository = userGroupAdminRepository;
     this.userGroupPermissionRepository = userGroupPermissionRepository;
   }
 
   public void testGet(
       int port, UserGroupAuthorisedActivityType activity, BundleUrlGetter bundleUrlGetter) {
-    setUpTestUserPermission(activity);
+    if (activity != null) {
+      setUpTestUserPermission(activity);
+    }
+
     BundleOfUsefulTestStuff bundle = getTestBundle();
 
     String url = String.format("http://localhost:%d/api/%s", port, bundleUrlGetter.getUrl(bundle));
     ResponseEntity<String> response = restTemplate.getForEntity(url, String.class);
     assertThat(response.getStatusCode()).as("GET is OK").isEqualTo(HttpStatus.OK);
 
-    deleteAllPermissions();
-    restoreDummyUserAndOtherGubbins(bundle); // Restore the user etc so that user tests still work
+    if (activity != null) {
+      deleteAllPermissions();
+      restoreDummyUserAndOtherGubbins(bundle); // Restore the user etc so that user tests still work
 
-    try {
-      restTemplate.getForEntity(url, String.class);
-      fail("GET API call was not forbidden, but should have been");
-    } catch (HttpClientErrorException expectedException) {
-      assertThat(expectedException.getStatusCode())
-          .as("GET is FORBIDDEN")
-          .isEqualTo(HttpStatus.FORBIDDEN);
+      try {
+        restTemplate.getForEntity(url, String.class);
+        fail("GET API call was not forbidden, but should have been");
+      } catch (HttpClientErrorException expectedException) {
+        assertThat(expectedException.getStatusCode())
+            .as("GET is FORBIDDEN")
+            .isEqualTo(HttpStatus.FORBIDDEN);
+      }
     }
   }
 
@@ -192,6 +202,7 @@ public class IntegrationTestHelper {
     UserGroup group = setupDummyGroup(UUID.randomUUID());
     UserGroup secondGroup = setupDummyGroup(UUID.randomUUID());
     UserGroupMember userGroupMember = setupDummyGroupMember(UUID.randomUUID(), user, group);
+    UserGroupAdmin userGroupAdmin = setupDummyGroupAdmin(UUID.randomUUID(), user, group);
     UserGroupPermission userGroupPermission = setupDummyGroupPermission(UUID.randomUUID(), group);
 
     BundleOfUsefulTestStuff bundle = new BundleOfUsefulTestStuff();
@@ -204,6 +215,7 @@ public class IntegrationTestHelper {
     bundle.setUserId(user.getId());
     bundle.setGroupId(group.getId());
     bundle.setGroupMemberId(userGroupMember.getId());
+    bundle.setGroupAdminId(userGroupAdmin.getId());
     bundle.setGroupPermissionId(userGroupPermission.getId());
     bundle.setSecondGroupId(secondGroup.getId());
 
@@ -215,6 +227,7 @@ public class IntegrationTestHelper {
     UserGroup group = setupDummyGroup(bundle.getGroupId());
     setupDummyGroup(bundle.getSecondGroupId());
     setupDummyGroupMember(bundle.getGroupMemberId(), user, group);
+    setupDummyGroupAdmin(bundle.getGroupAdminId(), user, group);
     setupDummyGroupPermission(bundle.getGroupPermissionId(), group);
   }
 
@@ -243,6 +256,15 @@ public class IntegrationTestHelper {
     return groupMember;
   }
 
+  private UserGroupAdmin setupDummyGroupAdmin(UUID groupAdminId, User user, UserGroup group) {
+    UserGroupAdmin groupAdmin = new UserGroupAdmin();
+    groupAdmin.setId(groupAdminId);
+    groupAdmin.setGroup(group);
+    groupAdmin.setUser(user);
+    groupAdmin = userGroupAdminRepository.saveAndFlush(groupAdmin);
+    return groupAdmin;
+  }
+
   private UserGroupPermission setupDummyGroupPermission(UUID groupPermissionId, UserGroup group) {
     UserGroupPermission permission = new UserGroupPermission();
     permission.setId(groupPermissionId);
@@ -255,6 +277,7 @@ public class IntegrationTestHelper {
   private void deleteAllPermissions() {
     userGroupPermissionRepository.deleteAllInBatch();
     userGroupMemberRepository.deleteAllInBatch();
+    userGroupAdminRepository.deleteAllInBatch();
     userGroupRepository.deleteAllInBatch();
     userRepository.deleteAllInBatch();
   }
